Add observer on changes to the portfolio total
https://github.com/Alin-Barsan/Fidelity-Portfolio-Total-Chrome-Extension/issues/1

When the total was updated by navigating around (or just by it legitimately changing from portfolio value changing), our script didn't do anything because it was written to statically execute once.  

Added an observer to listen to changes to the total, and then recalculate and update it given those changes.  
Was careful to disable the observer when making those changes as we could trigger our own observer and crash the page with an infinite loop.